requirements.txt
opencv-contrib-python==4.8.0.76
pyserial==3.5
numpy==1.24.3
py-zabbix==1.1.7
